backend/requirements.txt
azure-functions==1.24.0
cffi==2.0.0
cryptography==46.0.5
MarkupSafe==3.0.3
pycparser==3.0
PyJWT==2.11.0
python-dotenv==1.2.1
Werkzeug==3.1.5
